make proper python package
