+++ cob_map_accessibility_analysis/ros/src/cob_map_accessibility_analysis/__init__.py
@@ -1,0 +1,1 @@
+
