+++ cob_map_accessibility_analysis/setup.py
@@ -1,0 +1,11 @@
+#!/usr/bin/env python
+
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+    packages=['cob_map_accessibility_analysis'],
+    package_dir={'': 'ros/src'}
+)
+
+setup(**d)
